+++ Day21 - Chronal Conversion/Device/Instruction.cs
@@ -2,7 +2,7 @@ using System.Text.RegularExpressions;
 
 namespace AdventOfCode.Year2018.Day21.Device;
 
-public readonly record struct Instruction(Opcode Opcode, byte A, byte B, byte C)
+public readonly record struct Instruction(Opcode Opcode, uint A, uint B, byte C)
 {
 	private static readonly Regex _regex = new(@"^[ \t]*([a-z]+) (\d+) (\d+) (\d+)[ \t]*$", RegexOptions.Compiled);
 
@@ -20,11 +20,11 @@ public readonly record struct Instruction(Opcode Opcode, byte A, byte B, byte C)
 			{
 				throw new FormatException($"Invalid opcode: '{match.Groups[1].ValueSpan}'.");
 			}
-			if (!byte.TryParse(match.Groups[2].ValueSpan, out byte a))
+			if (!uint.TryParse(match.Groups[2].ValueSpan, out uint a))
 			{
 				throw new FormatException($"Invalid instruction value A: '{match.Groups[2].ValueSpan}'.");
 			}
-			if (!byte.TryParse(match.Groups[3].ValueSpan, out byte b))
+			if (!uint.TryParse(match.Groups[3].ValueSpan, out uint b))
 			{
 				throw new FormatException($"Invalid instruction value B: '{match.Groups[3].ValueSpan}'.");
 			}
